+++ libc/stdlib/strtold.c
@@ -1,0 +1,288 @@
+/* Copyright (c) 2002-2005  Michael Stumpf  <mistumpf@de.pepperl-fuchs.com>
+   Copyright (c) 2006,2008  Dmitry Xmelkov
+   Copyright (c) 2025       Georg-Johann Lay  
+
+   All rights reserved.
+
+   Redistribution and use in source and binary forms, with or without
+   modification, are permitted provided that the following conditions are met:
+
+   * Redistributions of source code must retain the above copyright
+     notice, this list of conditions and the following disclaimer.
+   * Redistributions in binary form must reproduce the above copyright
+     notice, this list of conditions and the following disclaimer in
+     the documentation and/or other materials provided with the
+     distribution.
+   * Neither the name of the copyright holders nor the names of
+     contributors may be used to endorse or promote products derived
+     from this software without specific prior written permission.
+
+   THIS SOFTWARE IS PROVIDED BY THE COPYRIGHT HOLDERS AND CONTRIBUTORS "AS IS"
+   AND ANY EXPRESS OR IMPLIED WARRANTIES, INCLUDING, BUT NOT LIMITED TO, THE
+   IMPLIED WARRANTIES OF MERCHANTABILITY AND FITNESS FOR A PARTICULAR PURPOSE
+   ARE DISCLAIMED. IN NO EVENT SHALL THE COPYRIGHT OWNER OR CONTRIBUTORS BE
+   LIABLE FOR ANY DIRECT, INDIRECT, INCIDENTAL, SPECIAL, EXEMPLARY, OR
+   CONSEQUENTIAL DAMAGES (INCLUDING, BUT NOT LIMITED TO, PROCUREMENT OF
+   SUBSTITUTE GOODS OR SERVICES; LOSS OF USE, DATA, OR PROFITS; OR BUSINESS
+   INTERRUPTION) HOWEVER CAUSED AND ON ANY THEORY OF LIABILITY, WHETHER IN
+   CONTRACT, STRICT LIABILITY, OR TORT (INCLUDING NEGLIGENCE OR OTHERWISE)
+   ARISING IN ANY WAY OUT OF THE USE OF THIS SOFTWARE, EVEN IF ADVISED OF THE
+   POSSIBILITY OF SUCH DAMAGE. */
+
+#if !defined(__AVR_TINY__) && __SIZEOF_LONG_DOUBLE__ == 8
+
+#include <ctype.h>
+#include <errno.h>
+#include <limits.h>
+#include <math.h>
+#include <stdlib.h>
+#include <stdint.h>
+
+#include <avr/flash.h>
+#include <avr/pgmspace.h>
+#include <bits/attribs.h>
+
+#include "alias.h"
+#include "best_as.h"
+#include "sectionname.h"
+
+#define ARRAY_SIZE(X) (sizeof(X) / sizeof(*X))
+
+#define INFINITYl (__builtin_infl())
+#define NANl      (__builtin_nanl(""))
+
+extern int isfinitel (long double);
+
+// libgcc / LibF7
+extern long double __floatundidf (uint64_t);
+
+#define AS __BEST_AS
+
+// Like PSTR, but for AS.
+#define ASSTR(s) \
+    (__extension__({ static const AS char __c[] = (s); &__c[0]; }))
+
+#ifdef __AVR_HAVE_ELPM__
+// Don't assume we have __flashx.
+#define READ_LD(p) pgm_read_long_double_far ((uintptr24_t) (p))
+extern int strncasecmp_AS (const char*,
+			   const AS char*, size_t) __asm ("strncasecmp_PF");
+#else
+#define READ_LD(p) flash_read_long_double (p)
+#define strncasecmp_AS(a, b, c) strncasecmp_F (a, b, c)
+#endif // ELPM?
+
+#define PROG_SECTION __attribute__((__section__(".progmemx.data.pwr_10L")))
+
+PROG_SECTION static const AS long double pwr_p10L[] =
+{
+    1e+1L, 1e+2L, 1e+4L, 1e+8L, 1e+16L, 1e+32L, 1e+64L, 1e+128L, 1e+256L
+};
+
+PROG_SECTION static const AS long double pwr_m10L[] =
+{
+    1.0e-1L, 1e-2L, 1e-4L, 1e-8L, 1e-16L, 1e-32L, 1e-64L, 1e-128L, 1e-256L
+};
+
+
+#define N_PWR ARRAY_SIZE (pwr_p10L)
+
+
+static __ATTR_ALWAYS_INLINE__ int
+mulhi10 (int i)
+{
+    register int r24 __asm("24") = i;
+    __asm ("%~call __mulhi_const_10"
+           : "+r" (r24) :: "22", "23");
+    return r24;
+}
+
+// misc/madddi10.S: return 10 * a + b;
+extern uint64_t madddi10 (uint64_t a, uint8_t b) __asm("__madddi_const_10");
+
+
+/* PSTR() is not used to save 1 byte per string: '\0' at the tail.	*/
+static const AS char pstr_inf[3] = { 'I','N','F' };
+static const AS char pstr_inity[5] = { 'I','N','I','T','Y' };
+static const AS char pstr_nan[3] = { 'N','A','N' };
+
+/**  The strtold() function converts the initial portion of the string pointed
+     to by \a nptr to \c long \c double representation.
+
+     The expected form of the string is an optional plus ( \c '+' ) or minus
+     sign ( \c '-' ) followed by a sequence of digits optionally containing
+     a decimal-point character, optionally followed by an exponent.  An
+     exponent consists of an \c 'E' or \c 'e', followed by an optional plus
+     or minus sign, followed by a sequence of digits.
+
+     Leading white-space characters in the string are skipped.
+
+     The strtold() function returns the converted value, if any.
+
+     If \a endptr is not \c NULL, a pointer to the character after the last
+     character used in the conversion is stored in the location referenced by
+     \a endptr.
+
+     If no conversion is performed, zero is returned and the value of
+     \a nptr is stored in the location referenced by \a endptr.
+
+     If the correct value would cause overflow, plus or minus \c INFINITY is
+     returned (according to the sign of the value), and \c ERANGE is stored
+     in \c errno.  If the correct value would cause underflow, zero is
+     returned and \c ERANGE is stored in \c errno.
+ */
+ATTRIBUTE_CLIB_SECTION
+long double
+strtold (const char *nptr, char **endptr)
+{
+    union
+    {
+	uint64_t u64;
+	long double ldbl;
+    } x;
+    uint8_t c;
+    uint8_t flag = 0;
+
+#define FL_MINUS    0x01	// number is negative
+#define FL_ANY	    0x02	// any digit was read
+#define FL_OVFL	    0x04	// mantissa overflow: dropping excess precision
+#define FL_DOT	    0x08	// decimal '.' was seen
+#define FL_MEXP	    0x10	// exponent 'e' is neg
+
+    if (endptr)
+	*endptr = (char*) nptr;
+
+    do
+	c = *nptr++;
+    while (isspace (c));
+
+    if (c == '-')
+    {
+	flag = FL_MINUS;
+	c = *nptr++;
+    }
+    else if (c == '+')
+	c = *nptr++;
+
+    if (!strncasecmp_AS (nptr - 1, pstr_inf, 3))
+    {
+	nptr += 2;
+	if (!strncasecmp_AS (nptr, pstr_inity, 5))
+	    nptr += 5;
+	if (endptr)
+	    *endptr = (char*) nptr;
+	return flag & FL_MINUS ? -INFINITYl : +INFINITYl;
+    }
+
+    // NAN() construction is not realised.
+    // Length would be 3 characters only.
+    if (!strncasecmp_AS (nptr - 1, pstr_nan, 3))
+    {
+	if (endptr)
+	    *endptr = (char*) nptr + 2;
+	return NANl;
+    }
+
+    x.u64 = 0;
+    int expo = 0;
+
+    // Read the mantissa, keeping stock of dot's (if any) position in expo.
+    for (;;)
+    {
+	c -= '0';
+
+	if (c <= 9)
+	{
+	    flag |= FL_ANY;
+	    if (flag & FL_OVFL)
+	    {
+		if (!(flag & FL_DOT))
+		    expo += 1;
+	    }
+	    else
+	    {
+		if (flag & FL_DOT)
+		    expo -= 1;
+		// x.u64 = 10 * x.u64 + c
+		x.u64 = madddi10 (x.u64, c);
+		if (x.u64 >= (ULLONG_MAX - 9) / 10)
+		    flag |= FL_OVFL;
+	    }
+	}
+	else if (c == (('.'-'0') & 0xff) && !(flag & FL_DOT))
+	    flag |= FL_DOT;
+	else
+	    break;
+
+	c = *nptr++;
+    }
+
+    // Read the exponent (if any).
+    if (c == (('e'-'0') & 0xff) || c == (('E'-'0') & 0xff))
+    {
+	c = *nptr++;
+	int i = 2;
+	if (c == '-')
+	{
+	    flag |= FL_MEXP;
+	    c = *nptr++;
+	}
+	else if (c == '+')
+	    c = *nptr++;
+	else
+	    i = 1;
+	c -= '0';
+	if (c > 9)
+	    nptr -= i;
+	else
+	{
+	    i = 0;
+	    do
+	    {
+		if (i < 3200)
+		    i = mulhi10 (i) + c;	// i = 10*i + c
+		c = *nptr++ - '0';
+	    } while (c <= 9);
+	    if (flag & FL_MEXP)
+		i = -i;
+	    expo += i;
+	}
+    }
+
+    if ((flag & FL_ANY) && endptr)
+	*endptr = (char*) nptr - 1;
+
+    // Convert mantissa and sign to long double.
+    x.ldbl = __floatundidf (x.u64);
+    if ((flag & FL_MINUS) && (flag & FL_ANY))
+	x.ldbl = -x.ldbl;
+
+    // Apply the exponent using its binary expansion.
+    if (x.ldbl != 0)
+    {
+	const AS long double *l_pwr;
+
+	if (expo < 0)
+	{
+	    l_pwr = pwr_m10L + (N_PWR - 1);
+	    expo = -expo;
+	}
+	else
+	    l_pwr = pwr_p10L + (N_PWR - 1);
+
+	for (int pwr = 1 << (N_PWR - 1); pwr; pwr >>= 1, --l_pwr)
+	    for (; expo >= pwr; expo -= pwr)
+		x.ldbl *= READ_LD (l_pwr);
+
+	if (!isfinitel (x.ldbl) || x.ldbl == 0)
+	    errno = ERANGE;
+    }
+
+    return x.ldbl;
+}
+
+
+DALIAS_LD (strtold)
+double strtod (const char*, char**);
+
+#endif // !AVR_TINY && long double is 64-bit